x86: edison: enable I2C#6
Initial configuration came from the firmware. Which quite likely has
been used in the phones, where I2C #8, that is not part of Atom
peripheral, is in use. Thus we need to override the leftover.

In order to achieve this goal we make use of the new code in pinmux.c and
set the appropriate pin function.

From now on we can use I2C#6 in Linux kernel.

Signed-off-by: Georgii Staroselskii <georgii.staroselskii@emlid.com>